% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_path.R
\name{format_all_paths}
\alias{format_all_paths}
\title{Formats the file paths of all target function calls in the active document
using format_path()}
\usage{
format_all_paths(fns = c("here", "file.path"), message = TRUE)
}
\arguments{
\item{fns}{Target function calls to format. Defaults are "here" and
"file.path"}

\item{message}{Logical, specifying whether or not to print information about
the cleaning process (file being formatted and number of times each target
function call was found) to the console}
}
\description{
As an alternative to formatting paths one at a a time, this
  function allows all file paths in the active document to be formatted at
  once. By default, all here() and file.path() calls are targeted, but these
  can be removed and other user-defined targets added. The function outputs
  some information on the file being formatted and number of times each
  target function call was found.
}
