% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_clean_path.R
\name{here_clean_all}
\alias{here_clean_all}
\title{Clean paths in all here::here() and file.path() in the active document}
\usage{
here_clean_all(fns = c("here", "file.path"), message = TRUE)
}
\arguments{
\item{fns}{Function to}

\item{message}{Logical, specifying whether or not to print information about
the cleaning process (file being cleaned and number of here() and
file.path() calls)}
}
\description{
Clean paths in all here::here() and file.path() in the active document
}
