% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_clean_path.R
\name{split_path}
\alias{split_path}
\title{Split path by slashes and quote and comma-separate result.}
\usage{
split_path(text, fns = c("here", "file.path"))
}
\arguments{
\item{text}{File path to reformat}
}
\value{

}
\description{
Split path by slashes and quote and comma-separate result.
}
