% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_path.R
\name{get_info}
\alias{get_info}
\title{Get information on the highlighted terms from the editor}
\usage{
get_info(context)
}
\arguments{
\item{context}{Context of the text}
}
\value{
Information about the highlighted text
}
\description{
Get information on the highlighted terms from the editor
}
