% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_path.R
\name{here_clean_path}
\alias{here_clean_path}
\title{Correctly format path to match here::here() guidelines}
\usage{
here_clean_path()
}
\value{
Updates the highlighted path as recommended by here::here()
}
\description{
Correctly format path to match here::here() guidelines
}
