% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_path.R
\name{split_path}
\alias{split_path}
\title{Find and format paths}
\usage{
split_path(text, fns = c("here", "file.path"))
}
\arguments{
\item{text}{File path to reformat}

\item{fns}{Target function calls to clean. Defaults are "here" and
"file.path"}
}
\description{
Find and extract paths, split them by slashes, quote and
comma-separate the result, and reinsert into text.
}
\keyword{internal}
